Remove redundant casts (#13509)
Motivation:
We should remove unnecessary redundant type casing for primitive types
of byte, int and long.

Modification:
Removed unnecessary redundant type casing

Result:
No more redundant casting

+++ example/src/main/java/io/netty5/example/ocsp/OcspUtils.java
@@ -80,7 +80,7 @@ public final class OcspUtils {
         }
 
         byte[] encoded = taggedObject.getEncoded();
-        int length = (int) encoded[1] & 0xFF;
+        int length = encoded[1] & 0xFF;
         String uri = new String(encoded, 2, length, StandardCharsets.UTF_8);
         return URI.create(uri);
     }
